Move "new" and "save" operations for sparse package into own files and insert optional "ops" into the constructor so that e.g., media types can be overridden for sparse images
